♻️rename SettingsActivity and Fragment into UserPreferences_Activity
also move fragment into dedicated package

closes #224
