+++ DorisAndroid/src/main/java/fr/ffessm/doris/android/activities/DetailEntreeGlossaire_ElementViewActivity.java
@@ -258,7 +258,7 @@ public class DetailEntreeGlossaire_ElementViewActivity extends OrmLiteActionBarA
         // behavior of option menu
         int itemId = item.getItemId();
         if (itemId == R.id.detailentreeglossaire_elementview_action_preference) {
-            startActivity(new Intent(this, SettingsActivity.class));
+            startActivity(new Intent(this, UserPreferences_Activity.class));
             return true;
             //Start of user code additional menu action DetailEntreeGlossaire_ElementViewActivity
         } else if (itemId == R.id.detailentreeglossaire_elementview_action_aide) {
