+++ DorisAndroid/src/main/java/fr/ffessm/doris/android/activities/DetailsBibliographie_ElementViewActivity.java
@@ -263,7 +263,7 @@ public class DetailsBibliographie_ElementViewActivity extends OrmLiteActionBarAc
         // behavior of option menu
         int itemId = item.getItemId();
         if (itemId == R.id.detailsbibliographie_elementview_action_preference) {
-            startActivity(new Intent(this, SettingsActivity.class));
+            startActivity(new Intent(this, UserPreferences_Activity.class));
             return true;
             //Start of user code additional menu action DetailsBibliographie_ElementViewActivity
         } else if (itemId == R.id.detailsbibliographie_elementview_action_aide) {
