+++ DorisAndroid/src/main/java/fr/ffessm/doris/android/activities/DetailsParticipant_ElementViewActivity.java
@@ -279,7 +279,7 @@ public class DetailsParticipant_ElementViewActivity extends OrmLiteActionBarActi
         // behavior of option menu
         int itemId = item.getItemId();
         if (itemId == R.id.detailsparticipant_elementview_action_preference) {
-            startActivity(new Intent(this, SettingsActivity.class));
+            startActivity(new Intent(this, UserPreferences_Activity.class));
             return true;
             //Start of user code additional menu action DetailsParticipant_ElementViewActivity
         } else if (itemId == R.id.detailsparticipant_elementview_action_aide) {
