+++ DorisAndroid/src/main/java/fr/ffessm/doris/android/activities/Glossaire_ClassListViewActivity.java
@@ -260,7 +260,7 @@ public class Glossaire_ClassListViewActivity extends OrmLiteActionBarActivity<Or
         // behavior of option menu
         int itemId = item.getItemId();
         if (itemId == R.id.glossaire_classlistview_action_preference) {
-            startActivity(new Intent(this, SettingsActivity.class));
+            startActivity(new Intent(this, UserPreferences_Activity.class));
             return true;
             //Start of user code additional menu action Glossaire_ClassListViewActivity
