+++ DorisAndroid/src/main/java/fr/ffessm/doris/android/activities/ImagePleinEcran_CustomViewActivity.java
@@ -55,7 +55,7 @@ import android.os.Bundle;
 import android.os.Handler;
 import androidx.core.app.NavUtils;
 import androidx.core.app.TaskStackBuilder;
-import androidx.appcompat.app.ActionBar;
+
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
@@ -359,7 +359,7 @@ public class ImagePleinEcran_CustomViewActivity extends OrmLiteActionBarActivity
 
         int itemId = item.getItemId();
         if (itemId == R.id.imagepleinecran_customview_action_preference) {
-            startActivity(new Intent(this, SettingsActivity.class));
+            startActivity(new Intent(this, UserPreferences_Activity.class));
             return true;
             //Start of user code additional menu action ImagePleinEcran_CustomViewActivity
