+++ DorisAndroid/src/main/java/fr/ffessm/doris/android/activities/ListeBibliographieAvecFiltre_ClassListViewActivity.java
@@ -259,7 +259,7 @@ public class ListeBibliographieAvecFiltre_ClassListViewActivity extends OrmLiteA
         // behavior of option menu
         int itemId = item.getItemId();
         if (itemId == R.id.listebibliographieavecfiltre_classlistview_action_preference) {
-            startActivity(new Intent(this, SettingsActivity.class));
+            startActivity(new Intent(this, UserPreferences_Activity.class));
             return true;
             //Start of user code additional menu action ListeBibliographieAvecFiltre_ClassListViewActivity
