+++ DorisAndroid/src/main/java/fr/ffessm/doris/android/activities/ListeParticipantAvecFiltre_ClassListViewActivity.java
@@ -273,7 +273,7 @@ public class ListeParticipantAvecFiltre_ClassListViewActivity extends OrmLiteAct
         // behavior of option menu
         int itemId = item.getItemId();
         if (itemId == R.id.listeparticipantavecfiltre_classlistview_action_preference) {
-            startActivity(new Intent(this, SettingsActivity.class));
+            startActivity(new Intent(this, UserPreferences_Activity.class));
             return true;
             //Start of user code additional menu action ListeParticipantAvecFiltre_ClassListViewActivity
