+++ DorisAndroid/src/main/java/fr/ffessm/doris/android/activities/SplashScreen_CustomViewActivity.java
@@ -71,7 +71,6 @@ import android.view.View;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
-import androidx.activity.EdgeToEdge;
 import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowCompat;
@@ -300,7 +299,7 @@ public class SplashScreen_CustomViewActivity extends OrmLiteActionBarActivity<Or
     public boolean onOptionsItemSelected(MenuItem item) {
         // behavior of option menu
         if (item.getItemId() == R.id.splashscreen_customview_action_preference) {
-            startActivity(new Intent(this, SettingsActivity.class));
+            startActivity(new Intent(this, UserPreferences_Activity.class));
             return true;
             //Start of user code additional menu action SplashScreen_CustomViewActivity
